requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.23
requests>=2.31
psycopg2-binary>=2.9
python-dotenv>=1.0
pydantic>=1.10,<3
ratelimit>=2.2
celery>=5.3
redis>=5.0
